github clients: always collect metrics about rate limits
