+++ internal/extsvc/github/metrics.go
@@ -1,0 +1,18 @@
+package github
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+)
+
+var githubRemainingGauge = promauto.NewGaugeVec(prometheus.GaugeOpts{
+	// TODO: New name?!
+	Name: "TODO_src_github_rate_limit_remaining_v2",
+	Help: "Number of calls to GitHub's API remaining before hitting the rate limit.",
+}, []string{"resource", "name"})
+
+var githubRatelimitWaitCounter = promauto.NewCounterVec(prometheus.CounterOpts{
+	// TODO: New name?!
+	Name: "TODO_src_github_rate_limit_wait_duration_seconds",
+	Help: "The amount of time spent waiting on the rate limit",
+}, []string{"resource", "name"})
